requirements.txt
openai==1.3.5
python-dotenv==1.0.0
youtube_transcript_api==0.6.1
google-api-python-client==2.88.0



